requirements.txt
# Core AI Framework (Stable 0.2.x series)
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.43

# Google Gemini Integration
langchain-google-genai==1.0.8

# Document Processing & RAG
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0
pypdf==4.3.1

# Web & API Tools
streamlit==1.32.0
requests==2.31.0

# Database & Environment
mysql-connector-python==8.0.33
python-dotenv==1.0.1

# Additional for Weather
# requests is already included above
# pip install -r requirements.txt